Move RemoveGetItemPass (remove_getitem_op.py) to backends.transform
Multiple backends are using this pass - Arm, Neutron, XNNPACK.
