+++ backends/xnnpack/test/passes/test_remove_get_item_pass.py
@@ -7,7 +7,7 @@
 import unittest
 
 import torch
-from executorch.backends.xnnpack._passes.remove_getitem_op import RemoveGetItemPass
+from executorch.backends.transforms.remove_getitem_op import RemoveGetItemPass
 from executorch.backends.xnnpack.test.tester import RunPasses, Tester
